feat(card): create high-level CardWithUser component
