+++ packages/clay-sticker/src/index.tsx
@@ -7,7 +7,7 @@
 import classNames from 'classnames';
 import React from 'react';
 
-type DisplayType =
+export type DisplayType =
 	| 'danger'
 	| 'dark'
 	| 'info'
@@ -60,7 +60,7 @@ interface IProps extends React.HTMLAttributes<HTMLSpanElement> {
 const ClaySticker: React.FunctionComponent<IProps> = ({
 	children,
 	className,
-	displayType = 'primary',
+	displayType,
 	inline,
 	outside = false,
 	position,
